Fix response parse issues and add more comprehensive parse testing example
